refactor(package): Lotto, WinningNumbers, WinningCriteria를 domain 패키지로 이동
