+++ src/main/java/lotto/LottoNumberGenerator.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.Lotto;
+
 public interface LottoNumberGenerator {
     public Lotto create();
 }
