+++ src/main/java/lotto/domain/WinningCriteria.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public enum WinningCriteria {
     FIRST(6, false, 2_000_000_000L),
