+++ src/main/java/lotto/service/LottoService.java
@@ -2,7 +2,7 @@ package lotto.service;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.Lotto;
+import lotto.domain.Lotto;
 import lotto.LottoNumberGenerator;
 
 public class LottoService {
